Tweak default gdalwarp memory settings after NCI tests.
500MB was overkill, and produces similar procesing times to 250MB.

Changing the warp memory size (wm) altered the output image, so it has been removed. We don't know if the difference was better or worse (yet), but don't wish to alter output without confirmation from the science team.

Compared with the original result, our output images are half the size (due to better lossless compression), and take 8.7 seconds instead of two minutes to produce.